% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{o311_discovery}
\alias{o311_discovery}
\title{API discovery}
\usage{
o311_discovery()
}
\value{
A list containing details on the given open311 API.
}
\description{
Retrieve discovery information about the mounted endpoint.
}
\examples{
o311_api("zurich")

can_connect <- o311_ok()
if (can_connect) {
  o311_discovery()
}
}
