% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{o311_query}
\alias{o311_query}
\title{Query an open311 service}
\usage{
o311_query(path, ..., simplify = TRUE)
}
\arguments{
\item{path}{Path appendix used to access endpoint-specific operations.}

\item{...}{Additional query parameters.}

\item{simplify}{Whether to simplify the output using
\code{jsonlite::toJSON(..., simplify = TRUE)}.}
}
\value{
The parsed query output, either as a list or dataframe.
}
\description{
Make a generic open311 query. Some open311 implementations support
unique operations that are not included in the official documentation.
This function can be used to access these URL paths.
}
\examples{
# example code

}
