% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/requests.R
\name{o311_requests}
\alias{o311_requests}
\title{Query the current status of multiple requests}
\usage{
o311_requests(
  jurisdiction_id = NULL,
  service_code = NULL,
  start_date = NULL,
  end_date = NULL,
  status = NULL,
  ...
)
}
\arguments{
\item{jurisdiction_id}{ID of a specific jurisdiction. Only required if
an endpoint serves multiple jurisdictions.}

\item{service_code}{\code{[character]}

IDs of the service types to be queried. Defaults to all available codes of
an endpoint. A list of all available service codes can be retrieved using
\code{\link{o311_services}}.}

\item{start_date, end_date}{\code{[POSIXct]}

Start date and end date of the query results. Must be date-time objects.
If not specified, defaults to the last 90 days.}

\item{status}{\code{[character]}

Status of the public service ticket. Can be one of \code{"open"} or
\code{"closed"}. If \code{NULL}, returns all types of tickets.}

\item{...}{Further endpoint-specific parameters as documented in the
respective endpoint reference.}
}
\description{
Get data from a registered open311 endpoint.
}
