% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/services.R
\name{o311_services}
\alias{o311_services}
\title{Get service list}
\usage{
o311_services()
}
\value{
A dataframe containing information about each service. The following
columns are included:

\describe{

}
}
\description{
Get a list of available services. Services are unique to the endpoint / city
and thus require an attached jurisdiction using
\code{\link{o311_jurisdiction}}.
}
